Intern rpc/event name strings we keep around
In order to do CallResult rpc name resolution we are now doing a
simple string interning of rpc names (and event names) when we keep
them in the converstation data. This will help avoid an unbounded
string proliferation when handling Call/CallResult messages.

Note that at the moment the interning map
(jbl_converstation_data_t.strings) is a per-conversation map, but
really it should be a per-file map (not sure how to properly hook into
the dissector file-level init yet).