Resolve RPC names in CallResult / Yield messages
Keep track of RPC names used in Call and Invoke messages and use
them in CallResult / Yield messages in place of the raw sequence
numbers.

Add a preference to be able to turn this behavior off in case it
turns out to be using too much memory on long captures.